🐛 [nox] Fix patching of pre-commit hooks on Git Bash for Windows
The patched pre-commit hooks installed by the `pre-commit` Nox session modify
the `PATH` environment variable, using `os.pathsep` to separate prepended paths.
On Git Bash for Windows, `os.pathsep` is a shell-special character and needs to
be quoted. This commit quotes the path separator.

This fixes the following error on Windows when invoking pre-commit from Git:

```
.git/hooks/pre-commit: line 3: C:\Users\User\hypermodern-python\.nox\pre-commit\Scripts: Is a directory
.git/hooks/pre-commit: line 13: dirname: command not found
An error has occurred: FatalError: git failed. Is it installed, and are you in a Git repository directory?
```

Retrocookie-Original-Commit: cjolowicz/cookiecutter-hypermodern-python-instance@2d133a0